Moving utils from pkg/ibm to general
